--- src/ibex_bluesky_core/fitting/__init__.py
@@ -1,1 +1,0 @@
-"""For IBEX Bluesky scan fitting."""
